refactor(code): minimize plugin auto-update diff
