+++ libs/code/deepagents_code/_env_vars.py
@@ -356,11 +356,7 @@ user-supplied key is always preserved.
 """
 
 PLUGIN_AUTO_UPDATE = "DEEPAGENTS_CODE_PLUGIN_AUTO_UPDATE"
-"""Toggle background updates for installed marketplace plugins.
-
-Disabled by default. Set to a truthy value (`1`, `true`, `yes`, or `on`) or use
-`[plugins].auto_update` in config.toml to enable it.
-"""
+"""Toggle background updates for installed marketplace plugins."""
 
 PLUGIN_CACHE_DIR = "DEEPAGENTS_CODE_PLUGIN_CACHE_DIR"
 """Override the plugin install/marketplace cache root.
